+++ elfi/methods/parameter_inference.py
@@ -27,9 +27,10 @@ from elfi.methods.bo.gpy_regression import GPyRegression
 from elfi.methods.bo.utils import stochastic_optimization
 from elfi.methods.posteriors import BolfiPosterior, RomcPosterior
 from elfi.methods.results import BolfiSample, OptimizationResult, RomcSample, Sample, SmcSample
-from elfi.methods.utils import (GMDistribution, ModelPrior, NDimBoundingBox,
-                                arr2d_to_batch, batch_to_arr2d, ceil_to_batch_size,
-                                compute_ess, flat_array_to_dict, weighted_var)
+from elfi.methods.utils import (DensityRatioEstimation, GMDistribution, ModelPrior,
+                                NDimBoundingBox, arr2d_to_batch, batch_to_arr2d,
+                                ceil_to_batch_size, compute_ess, flat_array_to_dict,
+                                weighted_sample_quantile, weighted_var)
 from elfi.model.elfi_model import AdaptiveDistance, ComputationContext, ElfiModel, NodeReference
 from elfi.utils import is_array
 from elfi.visualization.visualization import ProgressBar
@@ -413,7 +414,7 @@ class Sampler(ParameterInference):
 
         """
         bar = kwargs.pop('bar', True)
-
+        self.bar = bar
         return self.infer(n_samples, *args, bar=bar, **kwargs)
 
     def _extract_result_kwargs(self):
@@ -719,13 +720,63 @@ class SMC(Sampler):
         self._rejection = None
         self._round_random_state = None
 
-    def set_objective(self, n_samples, thresholds):
-        """Set the objective of the inference."""
+    def set_objective(self, n_samples, thresholds=None, max_iter=10,
+                      adaptive_threshold=False, initial_quantile=0.20,
+                      q_threshold=0.99, densratio_estimation=None):
+        """Set objective for ABC-SMC inference.
+
+        Parameters
+        ----------
+        n_samples : int
+            Number of samples to generate
+        thresholds : list, optional
+            List of thresholds for ABC-SMC
+        max_iter : int, optional
+            Maximum number of iterations
+        adaptive_threshold : bool, optional
+            Boolean to indicate whether to use adaptive threshold selection
+        initial_quantile : float, optional
+            Initial selection quantile for the first round of adaptive-ABC-SMC
+        q_threshold : float, optional
+            Termination criteratia for adaptive-ABC-SMC
+        densratio_estimation : DensityRatioEstimation, optional
+
+        References
+        ----------
+        .. [1] Simola, U., Cisewski-Kehe, J., Gutmann, M.U. and Corander, J. 
+           "Adaptive Approximate Bayesian Computation Tolerance Selection."
+            Bayesian Analysis 1:1-27, 2021. https://doi.org/10.1214/20-BA1211
+
+        """
+        self.adaptive_quantile = initial_quantile
+        self.adaptive_threshold_value = np.Inf
+
+        if thresholds is None:
+            self.q_threshold = q_threshold
+            adaptive_threshold = True
+            rounds = max_iter - 1
+        else:
+            self.q_threshold = 1.0
+            adaptive_threshold = False
+            rounds = len(thresholds) - 1
+
+        if adaptive_threshold:
+            if densratio_estimation is None:
+                self.densratio = DensityRatioEstimation(n=100,
+                                                        epsilon=0.001,
+                                                        max_iter=200,
+                                                        abs_tol=0.01,
+                                                        fold=5,
+                                                        optimize=False)
+            else:
+                self.densratio = densratio_estimation
+
+        self.adaptive_threshold = adaptive_threshold
         self.objective.update(
             dict(
                 n_samples=n_samples,
                 n_batches=self.max_parallel_batches,
-                round=len(thresholds) - 1,
+                round=rounds,
                 thresholds=thresholds))
         self._init_new_round()
 
@@ -763,11 +814,20 @@ class SMC(Sampler):
 
         if self._rejection.finished:
             self.batches.cancel_pending()
+            if self.bar:
+                self.progress_bar.update_progressbar(self.progress_bar.scaling + 1,
+                                                     self.progress_bar.scaling + 1)
+            self._new_population = self._extract_population()
             if self.state['round'] < self.objective['round']:
-                self._populations.append(self._extract_population())
-                self.state['round'] += 1
-                self._init_new_round()
 
+                if self.adaptive_threshold:
+                    self._set_adaptive_quantile()
+                    self._set_adaptive_threshold()
+
+                if (self.adaptive_quantile * (self.state['round'] > 0)) < self.q_threshold:
+                    self._populations.append(self._new_population)
+                    self.state['round'] += 1
+                    self._init_new_round()
         self._update_objective()
 
     def prepare_new_batch(self, batch_index):
@@ -800,17 +860,17 @@ class SMC(Sampler):
     def _init_new_round(self):
         round = self.state['round']
 
-        reinit_msg = 'ABC-SMC Round {0} / {1}'.format(
-            round + 1, self.objective['round'] + 1)
-        self.progress_bar.reinit_progressbar(
-            scaling=(self.state['n_batches']), reinit_msg=reinit_msg)
+        if self.bar:
+            reinit_msg = 'ABC-SMC Round {0} / {1}'.format(
+                round + 1, self.objective['round'] + 1)
+            self.progress_bar.reinit_progressbar(
+                scaling=(self.state['n_batches']), reinit_msg=reinit_msg)
         dashes = '-' * 16
         logger.info('%s Starting round %d %s' % (dashes, round, dashes))
 
         # Get a subseed for this round for ensuring consistent results for the round
         seed = self.seed if round == 0 else get_sub_seed(self.seed, round)
         self._round_random_state = np.random.RandomState(seed)
-
         self._rejection = Rejection(
             self.model,
             discrepancy_name=self.discrepancy_name,
@@ -819,21 +879,25 @@ class SMC(Sampler):
             seed=seed,
             max_parallel_batches=self.max_parallel_batches)
 
-        self._rejection.set_objective(
-            self.objective['n_samples'], threshold=self.current_population_threshold)
+        if self.adaptive_threshold and self.state['round'] == 0:
+            self._rejection.set_objective(
+                self.objective['n_samples'], quantile=self.adaptive_quantile)
+        else:
+            self._rejection.set_objective(
+                self.objective['n_samples'], threshold=self.current_population_threshold)
 
     def _extract_population(self):
         sample = self._rejection.extract_result()
         # Append the sample object
         sample.method_name = "Rejection within SMC-ABC"
-        w, cov = self._compute_weights_and_cov(sample)
+        means, w, cov = self._compute_weights_means_and_cov(sample)
+        sample.means = means
         sample.weights = w
         sample.meta['cov'] = cov
         return sample
 
-    def _compute_weights_and_cov(self, pop):
-        params = np.column_stack(
-            tuple([pop.outputs[p] for p in self.parameter_names]))
+    def _compute_weights_means_and_cov(self, pop):
+        params = np.column_stack(tuple([pop.outputs[p] for p in self.parameter_names]))
 
         if self._populations:
             q_logpdf = GMDistribution.logpdf(params, *self._gm_params)
@@ -841,6 +905,8 @@ class SMC(Sampler):
             w = np.exp(p_logpdf - q_logpdf)
         else:
             w = np.ones(pop.n_samples)
+
+        means = params.copy()
 
         if np.count_nonzero(w) == 0:
             raise RuntimeError("All sample weights are zero. If you are using a prior "
@@ -856,7 +922,7 @@ class SMC(Sampler):
                            "Falling back to using unit covariance.")
             cov = np.diag(np.ones(params.shape[1]))
 
-        return w, cov
+        return means, w, cov
 
     def _update_objective(self):
         """Update the objective n_batches."""
@@ -867,13 +933,56 @@ class SMC(Sampler):
     @property
     def _gm_params(self):
         sample = self._populations[-1]
-        params = sample.samples_array
-        return params, sample.cov, sample.weights
+        return sample.means, sample.cov, sample.weights
 
     @property
     def current_population_threshold(self):
         """Return the threshold for current population."""
+        if self.adaptive_threshold:
+            return self.adaptive_threshold_value
+
         return self.objective['thresholds'][self.state['round']]
+
+    def _set_adaptive_threshold(self):
+        self.adaptive_threshold_value = weighted_sample_quantile(
+            x=self._new_population.discrepancies,
+            alpha=self.adaptive_quantile,
+            weights=self._new_population.weights)
+        logger.info('ABC-SMC: Estimated adaptive threshold %.3f' % (self.adaptive_threshold_value))
+
+    def _set_adaptive_quantile(self):
+        """Set adaptively the new threshold for current population."""
+        logger.info("ABC-SMC: Adapting quantile threshold...")
+        sample_tm0 = self._new_population
+        weights_tm0 = sample_tm0.weights
+        n_tm0 = weights_tm0.shape[0]
+        x_tm0 = sample_tm0.samples_array
+        sample_sigma_tm0 = np.sqrt(np.diag(sample_tm0.cov))
+        if self.state['round'] == 0:
+            x_tm1 = self._prior.rvs(size=n_tm0, random_state=self._round_random_state)
+            weights_tm1 = np.ones(n_tm0)
+            sample_sigma_tm1 = np.diag(np.atleast_2d(np.cov(x_tm1.reshape(n_tm0, -1),
+                                                            rowvar=False)))
+        else:
+            sample_tm1 = self._populations[-1]
+            x_tm1 = sample_tm1.samples_array
+            weights_tm1 = sample_tm1.weights
+            sample_sigma_tm1 = np.sqrt(np.diag(sample_tm1.cov))
+
+        if self.densratio.optimize:
+            sigma = list(10.0 ** np.arange(-1, 6))
+        else:
+            sigma_tm0 = np.max(sample_sigma_tm0)
+            sigma_tm1 = np.max(sample_sigma_tm1)
+            sigma = sigma_tm0 * sigma_tm1 / np.sqrt(np.abs(sigma_tm0 ** 2 - sigma_tm1 ** 2))
+        self.densratio.fit(x_tm0, x_tm1,
+                           weights_x=weights_tm0,
+                           weights_y=weights_tm1,
+                           sigma=sigma)
+        max_value = self.densratio.max_ratio()
+        max_value = 1.0 if max_value < 1.0 else max_value
+        self.adaptive_quantile = max(1 / max_value, 0.05)
+        logger.info('ABC-SMC: Estimated maximum density ratio %.5f' % (max_value))
 
 
 class AdaptiveDistanceSMC(SMC):
@@ -947,6 +1056,31 @@ class AdaptiveDistanceSMC(SMC):
         self.quantile = quantile
         self._init_new_round()
 
+    def update(self, batch, batch_index):
+        """Update the inference state with a new batch.
+
+        Parameters
+        ----------
+        batch : dict
+            dict with `self.outputs` as keys and the corresponding outputs for the batch
+            as values
+        batch_index : int
+
+        """
+        super(SMC, self).update(batch, batch_index)
+        self._rejection.update(batch, batch_index)
+
+        if self._rejection.finished:
+            self.batches.cancel_pending()
+            self.progress_bar.update_progressbar(self.progress_bar.scaling + 1,
+                                                 self.progress_bar.scaling + 1)
+            if self.state['round'] < self.objective['round']:
+                self._populations.append(self._extract_population())
+                self.state['round'] += 1
+
+                self._init_new_round()
+        self._update_objective()
+
     def _init_new_round(self):
         round = self.state['round']
 
@@ -987,11 +1121,12 @@ class AdaptiveDistanceSMC(SMC):
         meta = rejection_sample.meta
         meta['threshold'] = max(outputs[self.discrepancy_name])
         meta['accept_rate'] = self.objective['n_samples']/meta['n_sim']
-        method_name = "Rejection within SMC-ABC"
+        method_name = "Rejection within adaptive distance SMC-ABC"
         sample = Sample(method_name, outputs, self.parameter_names, **meta)
 
         # Append the sample object
-        w, cov = self._compute_weights_and_cov(sample)
+        means, w, cov = self._compute_weights_means_and_cov(sample)
+        sample.means = means
         sample.weights = w
         sample.meta['cov'] = cov
         return sample