+++ CHANGELOG.rst
@@ -1,10 +1,12 @@
 Changelog
 =========
 
+
+- Add adaptive threshold selection method for ABC-SMC
 - Modify ProgressBar-functionality
 - Add constrains to ExpIntVar-acquisition so that no queries will be outside prior support
 - Add ABC-SMC with adaptive distance
-- Add robust optimized monte carlo method
+- Add Robust optimisation Monte Carlo method
 - Fix small issues in ABC-SMC which did not work in 1-dimensional problems or with output names
 - Update README.md